// src/user.services/user.service.interface.ts
export interface ValidateSessionReturn {
  username: string;
  wallet: WalletData;
  safeProperties: SafeProperties;
  playerStats: PlayerStats;
  energy: { energy: number, timeUntilNextRecharge: number | null };
  uuid: string;
  accessToken: string;
  refreshToken: string;
  message: string;
  success: string;
}
// ...
export interface AuthenticateReturn {
  username: string;
  wallet: WalletData;
  safeProperties: SafeProperties;
  playerStats: PlayerStats;
  energy?: { energy: number, timeUntilNextRecharge: number | null };
  uuid: string;
  refreshToken: string;
  accessToken: string;
  message: string;
  success: string;
}
// ...
export interface PlayerStats {
    level: number;
    playerExp: number;
    availStatPoints: number;
    rank: string;
    statPointsSaved: {
      mainVocalist: number;
      rapper: number;
      leadDancer: number;
      leadVocalist: number;
      mainDancer: number;
      visual: number;
      leadRapper: number;
    };
  }
// ...
export interface SafeProperties {
    signupDate: number;
  }
// ...
export interface UserData {
    properties: {
      signupDate: number
      accountType: "beats",
      userId: string;
      username: string;
      password: string;
      localWallet: string;
      localWalletKey: string;
      playerStats: PlayerStats;
      suspended: Suspended
      country: string;
      deviceId: string;
      inventorySize: any;
      smartWalletAddress: string;
      encryptedAccessKey: string;
      preferredServer: string;
    };
  }

  export interface PasskeyUserData extends UserData {
    properties: UserData["properties"] & {
      id: string;
      publicKey: Uint8Array;
      counter: number;
    };
  }
  /**
   * Represents data related to a user's wallet.
   *
   * @interface WalletData
   * @property {string} smartWalletAddress - The address of the smart wallet.
   * @property {string} beatsBalance - The balance in beats.
   * @property {string} gmrBalance - The balance in kmr.
   * @property {string} nativeBalance - The native balance.
   */
export interface WalletData {
    smartWalletAddress: string;
    beatsBalance: string;
    nativeBalance: string;
    apeBalance: string;
  }
// ...
export interface User {
  deviceId: string
  userName: string
  password: string | ""
  email?: string
}


export interface PasskeyUser extends User {
  id: string
  publicKey: Uint8Array
  counter: number
}


/**
 * Represents a user in the system.
 *
 * @interface PreRegisterUser
 * @property {string} userName - The username of the user.
 * @property {string} password - The password of the user.
 * @property {string} deviceId - The device identifier of the user.
 */
export interface PreRegisterUser {
  userName: string
  password: string
  email?: string
}

/**
 * Represents the token scheme containing both a refresh token and an access token.
 *
 * @interface TokenScheme
 * @property {string} refreshToken - The refresh token used for obtaining a new access token.
 * @property {string} accessToken - The access token used for authentication and authorization.
 * @property {string} username - The username associated with the tokens.
 */
export interface TokenScheme {
  refreshToken: string
  accessToken: string
  userName: string
}
// ...
export interface Suspended {
  until: number | null
  reason: string 
}